fixing checkOwnership
